+++ api/controller/project.controller.ts
@@ -16,7 +16,7 @@ const getProjects = catchAsync(
       Project.find({
         user: req.userId,
       }),
-      req.query
+      req.query,
     )
       .filter()
       .sort()
@@ -35,7 +35,7 @@ const getProjects = catchAsync(
         projects,
       },
     });
-  }
+  },
 );
 
 const getProject = catchAsync(
@@ -47,7 +47,7 @@ const getProject = catchAsync(
 
     if (!project) {
       return next(
-        new AppError(`No project found with the id ${req.params.id}`, 404)
+        new AppError(`No project found with the id ${req.params.id}`, 404),
       );
     }
 
@@ -57,7 +57,7 @@ const getProject = catchAsync(
         project,
       },
     });
-  }
+  },
 );
 
 const createProject = catchAsync(
@@ -71,8 +71,8 @@ const createProject = catchAsync(
       return next(
         new AppError(
           `Project with the name '${req.body.name}' already exists.`,
-          StatusCodes.CONFLICT
-        )
+          StatusCodes.CONFLICT,
+        ),
       );
 
     const project = await Project.create({
@@ -89,7 +89,7 @@ const createProject = catchAsync(
         project,
       },
     });
-  }
+  },
 );
 
 const deleteProject = catchAsync(
@@ -106,14 +106,14 @@ const deleteProject = catchAsync(
 
     if (!project) {
       return next(
-        new AppError(`No project found with the id ${req.params.id}`, 404)
+        new AppError(`No project found with the id ${req.params.id}`, 404),
       );
     }
 
     await createNotificationService({
       name: generateNotificationName(
         NotificationTypeEnum.PROJECT_DELETED,
-        project.name
+        project.name,
       )!,
       data: project.toObject(),
       value: project.id,
@@ -126,7 +126,7 @@ const deleteProject = catchAsync(
       message: "Project successfully deleted.",
       data: null,
     });
-  }
+  },
 );
 
 const updateProject = catchAsync(
@@ -139,8 +139,8 @@ const updateProject = catchAsync(
       return next(
         new AppError(
           `No project found with the id ${req.params.id}`,
-          StatusCodes.NOT_FOUND
-        )
+          StatusCodes.NOT_FOUND,
+        ),
       );
     }
 
@@ -158,14 +158,14 @@ const updateProject = catchAsync(
       return next(
         new AppError(
           `Project with the name '${req.body.name}' already exists.`,
-          StatusCodes.CONFLICT
-        )
+          StatusCodes.CONFLICT,
+        ),
       );
 
     await createNotificationService({
       name: generateNotificationName(
         NotificationTypeEnum.PROJECT_UPDATED,
-        project.name
+        project.name,
       )!,
       data: project.toObject(),
       value: project.id,
@@ -180,7 +180,7 @@ const updateProject = catchAsync(
       message: "Project successfully updated.",
       data: { project },
     });
-  }
+  },
 );
 
 const addProjectToFavorites = catchAsync(
@@ -191,7 +191,7 @@ const addProjectToFavorites = catchAsync(
     });
     if (!existingProject) {
       return next(
-        new AppError(`No project found with the id ${req.params.id}`, 404)
+        new AppError(`No project found with the id ${req.params.id}`, 404),
       );
     }
 
@@ -202,7 +202,7 @@ const addProjectToFavorites = catchAsync(
       },
       {
         favorite: true,
-      }
+      },
     );
 
     res.status(StatusCodes.OK).json({
@@ -210,7 +210,7 @@ const addProjectToFavorites = catchAsync(
       message: "Project successfully added to favorites.",
       data: { project },
     });
-  }
+  },
 );
 
 const removeProjectFromFavorites = catchAsync(
@@ -221,7 +221,7 @@ const removeProjectFromFavorites = catchAsync(
     });
     if (!existingProject) {
       return next(
-        new AppError(`No project found with the id ${req.params.id}`, 404)
+        new AppError(`No project found with the id ${req.params.id}`, 404),
       );
     }
 
@@ -232,14 +232,14 @@ const removeProjectFromFavorites = catchAsync(
       },
       {
         favorite: false,
-      }
+      },
     );
 
     res.status(204).json({
       status: "success",
       message: "Project successfully removed from favorites",
     });
-  }
+  },
 );
 
 export {